Refactor archive content reading within tests.
Added IArchiveContentReader to BaseTests. This removes the coupling
between tests and reading of xml files. Tests can now use ArchiveBuilder
to build XML as it fits, and dump the xml as a Stream into the subject
that are under test.

Tests that are skipped are not yet adapted to the new test regime.