WIP: test read from tar as filestream.
There are challenges with fetching the position from TarInputStream-object, because it reads in buffers. Therefore the position is not related to the entry.
The tar must be read in blocks of 512 bytes. If the size of TarHeaders are known, it would be possible to skip blocks containing headers, while only extracting blocks containing files. More testing needed.